requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pymupdf